Backend/requirements.txt
openai==1.3.0
google-generativeai==0.3.0
groq==0.31.1
python-dotenv==1.0.0
Pillow==10.0.0
pandas==2.0.0
audio-recorder-streamlit==0.0.10
SpeechRecognition==3.10.0
pydub==0.25.1